+++ CHANGES.md
@@ -17,6 +17,8 @@ All releases are available on [Anaconda.org](https://anaconda.org/conda-forge/ge
 
 - {gh}`1173` Fix double taxation of `sonstige_private_vorsorge_m` and add new input
   variable `einnahmen.renten.basisrente_m` for Basisrente / Rürup-Rente.
+- {gh}`1167` Implement 100k children exclusion for Grundsicherung im Alter (§43 SGB
+  XII). ({ghuser}`MImmesberger`)
 - {gh}`1164` Proper treatment of gemischte Bedarfsgemeinschaften (SGB II + SGB XII) via
   Vertikalmethode. ({ghuser}`MImmesberger`)
 - {gh}`1163` Implement Bedarfsanteilmethode for Bürgergeld/ALG2.